# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  profileImageUrl?: string | null;
}

export interface GetCurrentAuthUserResponse {
  user: AuthUser | null;
}

export interface ExchangeMobileAuthorizationCodeBody {
  code: string;
  code_verifier: string;
  redirect_uri: string;
  state: string;
  nonce?: string | null;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  token: string;
}

export interface LogoutMobileSessionResponse {
  success: boolean;
}

export interface SyncDataBody {
  notes: unknown[];
  tasks: unknown[];
  tags: unknown[];
}

export interface SyncDataResponse {
  notes: unknown[];
  tasks: unknown[];
  tags: unknown[];
  syncedAt: string | null;
}