# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /data/sync:
    get:
      operationId: getDataSync
      tags: [data]
      summary: Fetch user's synced data
      responses:
        "200":
          description: User data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncDataResponse"
    post:
      operationId: postDataSync
      tags: [data]
      summary: Save user's data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncDataBody"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncDataResponse"
components:
  schemas:
...
    SyncDataBody:
      type: object
      properties:
        notes:
          type: array
          items: {}
        tasks:
          type: array
          items: {}
        tags:
          type: array
          items: {}
      required:
        - notes
        - tasks
        - tags
    SyncDataResponse:
      type: object
      properties:
        notes:
          type: array
          items: {}
        tasks:
          type: array
          items: {}
        tags:
          type: array
          items: {}
        syncedAt:
          type: string
          nullable: true
      required:
        - notes
        - tasks
        - tags
        - syncedAt